+++ include/options/PathDependentOption.h
@@ -10,14 +10,14 @@ class PathDependentOption
     : public Option
 {
 public:
-            PathDependentOption (OptionNature* nature, int maturity, int obsFreq);
+            PathDependentOption (OptionNature* nature, size_t maturity, size_t obsFreq);
     virtual ~PathDependentOption();
 
             int    observationFrequency()                                     const;
     virtual double payoff              (const std::vector<double>& spot_path) const = 0;
 
 protected:
-    int _obsFreq;
+    size_t _obsFreq;
 };
 
 #endif // PathDependentOption_h
