+++ src/options/AmericanOption.cpp
@@ -1,6 +1,6 @@
 #include "AmericanOption.h"
 
-AmericanOption::AmericanOption(OptionNature* nature, int maturity, double strike)
+AmericanOption::AmericanOption(OptionNature* nature, size_t maturity, double strike)
     : VanillaOption(nature, maturity, strike)
 {
 }
