+++ src/options/CashOrNothingOption.cpp
@@ -1,7 +1,7 @@
 #include "CashOrNothingOption.h"
 
 
-CashOrNothingOption::CashOrNothingOption(OptionNature* nature, int maturity, double strike, double rebate)
+CashOrNothingOption::CashOrNothingOption(OptionNature* nature, size_t maturity, double strike, double rebate)
     : DigitalOption(nature, maturity, strike)
     , _rebate(rebate)
 {
